refactor: create a module for `E2eiEnrollment` and define it there
